HOLO-2 <fix: 빌드 에러 해결>
